# Partial in place Xor