@lemire 2015-12-03
@bpot  For extra credits, would you throw in at least one unit test so we know that the function is a bit sane with respect to what it outputs?