@bpot 2015-12-03
@lemire I think there are already a couple https://github.com/tgruben/roaring/blob/7747b7aa1bd4e8ed06edd304a59dfe5d67ae1cc9/roaring_test.go#L1066-L1067 https://github.com/tgruben/roaring/blob/7747b7aa1bd4e8ed06edd304a59dfe5d67ae1cc9/roaring_test.go#L1182-L1183

They were failing as I was developing and fixing bugs.

Happy to add more if those aren't sufficient.