@lemire 2015-12-03
@bpot  Ok. I'll merge. Sorry about that.